存在しないcolorschmeになっていたので修正 何重にもダウンロードしないように修正

+++ install-script/bash/common/vim.sh
@@ -1,16 +1,22 @@
 #!/bin/bash
 
-CUR_DIR=`pwd`
-
 ## vim
 sudo apt install vim
-rm ~/.vimrc
-ln -s $CUR_DIR/.vimrc ~/.vimrc
-git config --global core.editor vim
-mkdir ~/.vim
-mkdir ~/.vim/colors
-curl -O https://raw.githubusercontent.com/nanotech/jellybeans.vim/master/colors/jellybeans.vim > ~/.vim/colors
 
+if [ ! -e ~/.vimrc ]; then
+    ln -s $CUR_DIR/.vimrc ~/.vimrc
+    git config --global core.editor vim
+else
+    echo "aleady exist .vimrc"
+fi
+
+ls ~/.vim/colors >/dev/null 2>&1
+if [ ! $? = 0 ]; then
+    mkdir -p ~/.vim/colors
+    curl -O https://raw.githubusercontent.com/jonathanfilip/vim-lucius/master/colors/lucius.vim > ~/.vim/colors/lucius.vim
+else
+    echo "already installed vim color scheme"
+fi
